refactored components and mapUtil tests

+++ src/main/java/com/guihvicentini/keycloakconfigtool/services/AuthenticationFlowImportService.java
@@ -5,6 +5,7 @@ import com.guihvicentini.keycloakconfigtool.mappers.AuthenticationFlowMapper;
 import com.guihvicentini.keycloakconfigtool.models.*;
 import com.guihvicentini.keycloakconfigtool.utils.ListUtil;
 import lombok.extern.slf4j.Slf4j;
+import org.keycloak.representations.idm.AuthenticationExecutionInfoRepresentation;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -34,13 +35,13 @@ public class AuthenticationFlowImportService {
         var toBeDeleted = ListUtil.getMissingConfigElements(actual, target);
         var toBeUpdated = ListUtil.getNonEqualConfigsWithSameIdentifier(target, actual);
 
-        addFlows2(realm, toBeAdded);
-        deleteFlows2(realm, toBeDeleted);
-        updateFlows2(realm, toBeUpdated);
+        addFlows(realm, toBeAdded);
+        deleteFlows(realm, toBeDeleted);
+        updateFlows(realm, toBeUpdated);
 
     }
 
-    private void updateFlows2(String realm, List<AuthenticationFlow> flows) {
+    private void updateFlows(String realm, List<AuthenticationFlow> flows) {
         flows.forEach(flow -> updateFlow(realm, flow));
     }
 
@@ -49,15 +50,17 @@ public class AuthenticationFlowImportService {
         addFlow(realm, flow);
     }
 
-    private void deleteFlows2(String realm, List<AuthenticationFlow> flows) {
+    private void deleteFlows(String realm, List<AuthenticationFlow> flows) {
         flows.forEach(flow -> deleteFlow(realm, flow));
     }
 
     private void deleteFlow(String realm, AuthenticationFlow flow) {
-        resourceAdapter.deleteFlow(realm, flow.getAlias());
+        if(!flow.isBuiltIn()) {
+            resourceAdapter.deleteFlow(realm, flow.getAlias());
+        }
     }
 
-    private void addFlows2(String realm, List<AuthenticationFlow> authenticationFlows) {
+    private void addFlows(String realm, List<AuthenticationFlow> authenticationFlows) {
         authenticationFlows.forEach(authenticationFlow -> addFlow(realm, authenticationFlow));
     }
 
@@ -76,11 +79,24 @@ public class AuthenticationFlowImportService {
         });
     }
 
-    private void createExecution(String realm, String alias, FlowElement execution) {
+    private void createExecution(String realm, String flowAlias, FlowElement execution) {
         AuthenticationExecution authenticationExecution = (AuthenticationExecution) execution;
-        resourceAdapter.addExecution(realm, alias, authenticationExecution);
-        resourceAdapter.updateExecution(realm, alias, authenticationExecution);
+        resourceAdapter.addExecution(realm, flowAlias, authenticationExecution);
+        resourceAdapter.updateExecution(realm, flowAlias, authenticationExecution);
+        if (authenticationExecution.getConfig() != null) {
+            createOrUpdateAuthenticatorConfig(realm, flowAlias, authenticationExecution.getAlias(), authenticationExecution.getConfig());
+        }
+    }
 
+    private void createOrUpdateAuthenticatorConfig(String realm, String flowAlias, String executionAlias,
+                                                   AuthenticatorConfigConfig authenticatorConfig) {
+        AuthenticationExecutionInfoRepresentation execution = resourceAdapter.getAuthenticationExecutionByName(realm, flowAlias,
+                executionAlias);
+        if(execution.getAuthenticationConfig() == null) {
+            resourceAdapter.createNewExecutionConfig(realm, execution.getId(), flowMapper.mapToRepresentation(authenticatorConfig));
+        } else {
+            resourceAdapter.updateExecutionConfig(realm, execution.getAuthenticationConfig(), flowMapper.mapToRepresentation(authenticatorConfig));
+        }
     }
 
     private void createSubFlow(String realm, String alias, FlowElement subFlow) {
